BasicAuth/requirements.txt
fastapi==0.115.6
uvicorn==0.23.0
passlib==1.7.4
python-dotenv==1.0.0
jinja2==3.1.4
python-multipart==0.0.19